Move common taste API under user resource
